Initial commit
Worked on the code before committing for a little bit to decide if this project
was worth pursuing. I think it is so I'm gonna go ahead and put it online.

Currently implemented stuff some basic hit object stuff along with a simple
highway renderer based on the `BasicSpriteRenderer`, which will probably
have to change. If I understood camera and transformation matrix stuff
better I would probably use the 3D renderer (and I hope to switch to that soon
with some help).